Changed some ints to shorts.
Amity was having an error with the api because some of the database values were smallint and written in the api as int. I had to change them all to short to make them compatible.

+++ basketballGAME/basketballAPI/models/Team.cs
@@ -11,9 +11,9 @@ public partial class Team
 
     public string TeamAbbreviation { get; set; } = null!;
 
-    public int Wins { get; set; }
+    public short Wins { get; set; }
 
-    public int Losses { get; set; }
+    public short Losses { get; set; }
 
     public virtual ICollection<Game> GameTeamNoOneNavigations { get; set; } = new List<Game>();
